+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]## [0.5.4.6b0]
+## [Unreleased]## [0.5.4.6b1]
+## Added
+
+## [0.5.4.6b0]- 2023-01-10
 ## Added
 - ScenarioDbTable.fixNanNoneNull: convert more 'nan' values to None and thus to a correct NULL in DB
 - ScenarioDbManager supports PostgreSQL